+++ src/frontend/apps/standalone_site/src/features/HomePage/components/HomePage.spec.tsx
@@ -1,6 +1,7 @@
 import { screen } from '@testing-library/react';
 import { render } from 'lib-tests';
 import React from 'react';
+import ReactTestUtils from 'react-dom/test-utils';
 
 import HomePage from './HomePage';
 
@@ -14,6 +15,16 @@ describe('<HomePage />', () => {
     expect(screen.getByText(/My ContentsShuffle/i)).toBeInTheDocument();
     expect(
       screen.getByRole('link', { name: '› See Everything' }),
+    ).toBeInTheDocument();
+
+    const imageRef = screen.getByAltText('Homepage Banner');
+    ReactTestUtils.Simulate.load(imageRef);
+
+    expect(screen.getByText(/Learn freely/i)).toBeInTheDocument();
+    expect(
+      screen.getByText(
+        /Online courses to discover, learn, progress and succeed/i,
+      ),
     ).toBeInTheDocument();
   });
 });